extract changes tree container and use it in compare
